BAC-31 refactor: 프로젝트 파일 구조 정리
