+++ src/main/java/com/monsoon/seedflowplus/domain/account/entity/Role.java
@@ -1,4 +1,4 @@
-package com.monsoon.seedflowplus.domain.account;
+package com.monsoon.seedflowplus.domain.account.entity;
 
 public enum Role {
     CLIENT,
