+++ src/main/java/com/monsoon/seedflowplus/domain/document/request/entity/QuotationRequestStatus.java
@@ -1,4 +1,4 @@
-package com.monsoon.seedflowplus.domain.document.request;
+package com.monsoon.seedflowplus.domain.document.request.entity;
 
 public enum QuotationRequestStatus {
     PENDING, // 대기
