+++ src/main/java/com/monsoon/seedflowplus/domain/product/entity/ProductBookmark.java
@@ -1,7 +1,7 @@
 package com.monsoon.seedflowplus.domain.product.entity;
 
 import com.monsoon.seedflowplus.core.common.entity.BaseCreateEntity;
-import com.monsoon.seedflowplus.domain.account.User;
+import com.monsoon.seedflowplus.domain.account.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
